Do not reconcile if annotation present

+++ operators/multiclusterobservability/controllers/placementrule/placementrule_controller.go
@@ -7,6 +7,7 @@ import (
 	"context"
 	"errors"
 	"reflect"
+	"strings"
 	"sync"
 	"time"
 
@@ -46,10 +47,11 @@ import (
 )
 
 const (
-	ownerLabelKey             = "owner"
-	ownerLabelValue           = "multicluster-observability-operator"
-	managedClusterObsCertName = "observability-managed-cluster-certs"
-	nonOCP                    = "N/A"
+	ownerLabelKey                                  = "owner"
+	ownerLabelValue                                = "multicluster-observability-operator"
+	managedClusterObsCertName                      = "observability-managed-cluster-certs"
+	nonOCP                                         = "N/A"
+	disableAddonAutomaticInstallationAnnotationKey = "addon.open-cluster-management.io/disable-automatic-installation"
 )
 
 var (
@@ -114,6 +116,15 @@ func (r *PlacementRuleReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	if config.IsPaused(mco.GetAnnotations()) {
 		reqLogger.Info("MCO reconciliation is paused. Nothing more to do.")
 		return ctrl.Result{}, nil
+	}
+
+	// Do not reconcile objects if this instance of mch has the
+	// `disableAddonAutomaticInstallationAnnotationKey` annotation
+	if value, ok := mco.GetAnnotations()[disableAddonAutomaticInstallationAnnotationKey]; ok &&
+		strings.EqualFold(value, "true") {
+
+		reqLogger.Info("Cluster has disable addon automatic installation annotation. Skip addon deploy")
+		return reconcile.Result{}, nil
 	}
 
 	if !deleteAll && !mco.Spec.ObservabilityAddonSpec.EnableMetrics {
